+++ src/frontend/apps/impress/src/features/docs/doc-editor/components/BlockNoteSuggestionMenu.tsx
@@ -9,8 +9,11 @@ import {
 import React, { useMemo } from 'react';
 import { useTranslation } from 'react-i18next';
 
+import { useConfig } from '@/core';
+
 import { DocsBlockSchema } from '../types';
 
+import { modGetAISlashMenuItems } from './AI';
 import {
   getCalloutReactSlashMenuItems,
   getDividerReactSlashMenuItems,
@@ -20,21 +23,28 @@ export const BlockNoteSuggestionMenu = () => {
   const editor = useBlockNoteEditor<DocsBlockSchema>();
   const { t } = useTranslation();
   const basicBlocksName = useDictionary().slash_menu.page_break.group;
+  const { data: conf } = useConfig();
 
   const getSlashMenuItems = useMemo(() => {
-    return async (query: string) =>
-      Promise.resolve(
+    return async (query: string) => {
+      const getAISlashMenuItems = await modGetAISlashMenuItems;
+
+      return Promise.resolve(
         filterSuggestionItems(
           combineByGroup(
             getDefaultReactSlashMenuItems(editor),
             getPageBreakReactSlashMenuItems(editor),
             getCalloutReactSlashMenuItems(editor, t, basicBlocksName),
             getDividerReactSlashMenuItems(editor, t, basicBlocksName),
+            conf?.AI_FEATURE_ENABLED && getAISlashMenuItems
+              ? getAISlashMenuItems(editor)
+              : [],
           ),
           query,
         ),
       );
-  }, [basicBlocksName, editor, t]);
+    };
+  }, [basicBlocksName, editor, t, conf?.AI_FEATURE_ENABLED]);
 
   return (
     <SuggestionMenuController